Use one level of tabs instead of two for measurement details.
